Move `CommentBarConfiguration` child types to new file
